Making division of the module qtconsole
