+++ src/qturtle/qtconsole/__init__.py
@@ -1,0 +1,4 @@
+from . import styles
+from . import itranspyler_qt_consoleapp 
+from . import itranspyler_widget_qt_consoleapp 
+from . import transpyler_console
